+++ src/gotenberg_client/__init__.py
@@ -2,5 +2,9 @@
 #
 # SPDX-License-Identifier: MPL-2.0
 from gotenberg_client._client import GotenbergClient
+from gotenberg_client._errors import BaseClientError
+from gotenberg_client._errors import CannotExtractHereError
+from gotenberg_client.responses import SingleFileResponse
+from gotenberg_client.responses import ZipFileResponse
 
-__all__ = ["GotenbergClient"]
+__all__ = ["GotenbergClient", "SingleFileResponse", "ZipFileResponse", "BaseClientError", "CannotExtractHereError"]