+++ src/gotenberg_client/_merge.py
@@ -7,10 +7,11 @@ from typing import List
 from httpx import Client
 
 from gotenberg_client._base import BaseApi
-from gotenberg_client._base import BaseRoute
+from gotenberg_client._base import BaseZipFileResponseRoute
+from gotenberg_client._types import Self
 
 
-class MergeRoute(BaseRoute):
+class MergeRoute(BaseZipFileResponseRoute):
     """
     Handles the merging of a given set of files
     """
@@ -19,7 +20,7 @@ class MergeRoute(BaseRoute):
         super().__init__(client, api_route)
         self._next = 1
 
-    def merge(self, files: List[Path]) -> "MergeRoute":
+    def merge(self, files: List[Path]) -> Self:
         """
         Adds the given files into the file mapping.  This method will maintain the
         ordering of the list.  Calling this method multiple times may not merge